incorporate chart component
