+++ views/layouts/chart-wrapper.hbs
@@ -21,7 +21,7 @@
   </div>
   {{> footer}}
   <script>var exports = {};</script>
-  <script src="js/main.js"></script>
+  <script src="js/main.bundle.js"></script>
 </body>
 
 </html>
